refactor(spindle-ui): replace icon index imports with direct imports
